What is this?

Documentation for Interp1d.

Interp1d.jl

A Julia package for univariate interpolations.

This package is inspired by scipy.interpolate.interp1d function.

Requirements

Julia 1.6.x or higher.

Install

<!–

using Pkg;Pkg.add("Interp1d")

and then just import it with using Interp1d. –>

If you want use latest development version,

using Pkg;Pkg.add(url = "https://github.com/AtsushiSakai/Interp1d.jl.git")

How to use

You can use many kind of univariate interpolations:

using Interp1d
using PyPlot
xi = collect(0.0:10.0) # x of interpolated points
yi = exp.(-xi./3.0) # y of interpolated points
x = collect(-1.0:0.1:11.0) # sampling points
for mode in INTERP_MODE_LIST # Do interpolations for all modes.
    f = interp(xi, yi, mode) # get an interpolation function
    y = f.(x) # Do interpolation
    plot(x, y, label=string(mode))
end
plot(xi, yi, "xr")
legend()

<p align="left"><img src="assets/allmodeplot_sample.png" alt="drawing" width="500"/></p>

Details and other tips can be found in the latest documentaion:

Interp1d.jl Documentation

Author

Atsushi Sakai (@Atsushi_twi)

Interp1d.Nearest

Interp1d.Previous

Interp1d.interp

Interp1d.interp

Interp1d.Nearest — Type

Nearest interpolation type

Interp1d.Previous — Type

Previous interpolation type

Interp1d.interp — Method

interp(xi::Vector{T}, yi::VecOrMat{T}, ::Type{Nearest}; copy=true, sorted=false) where T

Returns a nearest interpolation function f(x) based on interpolated points xi and yi.

If xi has a duplicated value, ArgumentError is thrown.

If copy is true, this function makes internal copies of xi and yi, otherwise it uses references.

If sorted is false, this function sorts xi first, otherwise it assumes xi is monotonically increasing.

Examples

This is a simple interpolation example:

julia> using Interp1d

julia> xi = [-1.0, 0.0, 3.0/2.0];

julia> yi = [2.0, 1.0, 3.0];

julia> f = interp(xi, yi, Nearest);

julia> f(1.0)
3.0

yi can be matrix for multi univerate interpolations. The number of rows of yi must be equal to the length of xi.

julia> using Interp1d

julia> xi = [-1.0, 0.0, 3.0/2.0];

julia> yi = [ 2.0  1.0 3.0;
             -1.0 -4.4 2.0];

julia> f = interp(xi, yi, Nearest);

julia> f(0.0)
2-element Vector{Float64}:
  1.0
 -4.4

Interp1d.interp — Method

interp(xi::Vector{T}, yi::VecOrMat{T}, ::Type{Previous}; copy=true, sorted=false) where T

Returns a nearest interpolation function f(x) based on interpolated points xi and yi.

If xi has a duplicated value, ArgumentError is thrown.

If copy is true, this function makes internal copies of xi and yi, otherwise it uses references.

If sorted is false, this function sorts xi first, otherwise it assumes xi is monotonically increasing.